refactor: fixed rooms routing
