+++ expo/app/(tabs)/rooms/_layout.tsx
@@ -1,0 +1,5 @@
+import { Slot } from "expo-router";
+
+export default function RoomsTabLayout() {
+  return <Slot />;
+}
