+++ expo/app/(tabs)/rooms/index.tsx
@@ -3,7 +3,6 @@ import { Link } from "expo-router";
 import React from "react";
 import { Text, View } from "react-native";
 
-// TODO : review
 export default function TabsRooms() {
   return (
     <View>
